requirements.txt
python-telegram-bot==21.10
requests==2.31.0
python-dotenv==1.0.0
aiohttp>=3.11.0
google-genai
